lanner: create a medium quality version that uses a single material
The medium quality version uses the 'shiny' material everywhere and has no
undercarriage. This allows it to be a single static mesh that can be
rendered in a single batch, whereas the high LOD version has a bunch (~10)
of different components for the undercarriage, and three materials used
on the main hull (shiny top, matte bottom, and a few emissive parts for
the windows.

The poly count is only reduced slightly, but the reduction in render
batches should be worthwhile (perhaps, anyway).